@@ -10,14 +10,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/animations": "^14.1.3",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/forms": "^14.1.3",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
     "@fortawesome/angular-fontawesome": "^0.11.0",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-brands-svg-icons": "^6.1.1",
@@ -31,9 +31,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.3",
-    "@angular/cli": "~13.3.3",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular/cli": "^14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
