@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { faPenToSquare } from '@fortawesome/free-solid-svg-icons';
+import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
+import { faXmark } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'app-educacion',
@@ -6,7 +9,9 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./educacion.component.css']
 })
 export class EducacionComponent implements OnInit {
-
+  faPenToSquare = faPenToSquare;
+  faPlusCircle = faPlusCircle;
+  faXmark = faXmark;
   constructor() { }
 
   ngOnInit(): void {
