@@ -1,12 +1,16 @@
 import { Component, OnInit } from '@angular/core';
-
+import { faPenToSquare } from '@fortawesome/free-solid-svg-icons';
+import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
+import { faXmark } from '@fortawesome/free-solid-svg-icons';
 @Component({
   selector: 'app-experiencia',
   templateUrl: './experiencia.component.html',
   styleUrls: ['./experiencia.component.css']
 })
 export class ExperienciaComponent implements OnInit {
-
+  faPenToSquare = faPenToSquare;
+  faPlusCircle = faPlusCircle;
+  faXmark = faXmark;
   constructor() { }
 
   ngOnInit(): void {
